MDEV-32570 Prep: Split read_log_event into non-checksum version
Preparation for MDEV-32570. When fragmenting a large row event into
multiple smaller fragment events, each fragment event will have its own
checksum attached, thereby negating the need to also store the checksum
of the overall large event.

The existing code assumes all events will always have checksums, but
this won't be true for the rows events that are re-assembled on the
replicas. This patch prepares for this by splitting the logic which
reads in and creates Log_event objects into two pieces, one which
handles the checksum validation; and the other which reads the raw
event data (without the checksum) and creates the object.

All existing code is unchanged which uses the checksum-assuming version
of the event reader. MDEV-32570 will be the only case which will bypass
the checksum logic, and will directly create its rows log events from
memory without validating checksums (as the checksums will have already
been validated by each individual fragment event).

+++ sql/rpl_utility.cc
@@ -295,7 +295,7 @@ table_def::~table_def()
     event_buf argument is not a pointer to const.
 
 */
-bool event_checksum_test(uchar *event_buf, ulong event_len,
+bool event_checksum_test(uchar *event_buf, size_t event_len,
                          enum_binlog_checksum_alg alg)
 {
   bool res= FALSE;
